fix: docs version switcher [MLG-1524]
A bug in the pydata-sphinx-theme caused the version switcher to not
properly avoid 404 errors when switching between versions, if the
current doc url did not exist in the target version.

pydata-sphinx-theme has a fix in 0.14.0, but we cannot upgrade due to
sphinx-book-theme.  So I backported the relevant fix.  This commit
pulls the backported fix into the determined docs build system.