fix: improve trial log request cancelling
this fixes an issue where resetting the filter settings on trials with
logs that take a long time to respond (over 500 ms) can cause the catch
up log request to get cancelled. To fix, we skip aborting the request
when handling filter changes if the filter is being updated to an empty
object. This is because the request was already reset when the filter
was reset. We also ensure that old requests don't stick around by
merging the outer abort controller with the one that the stream config
gives us.