fix: remove possible hang from ship_logs.py [MLG-1565] (#8803)
Previously, after the main subprocess exited or was killed, the log
shipper would wait without limit for the stdout and stderr readers to
finish reading, under the assumption that after the main subprocess died
they would be quickly drained.

However, if the main subprocess spun off other, non-terminating child
processes, there would remain writers to the stdout and stderr, and this
assumption caused a hang.

I named this the "Orphaned Grandchild Problem", and it is concerning
because it can cause severe problems for users making fairly innocent
mistakes.  It didn't appear earlier because most training scripts that
we influence practice structured concurrency.

This fix includes a regression test.