chore: api's default retry now session's default retry (#8872)
Pre-patch, a `Session` that was constructed without a `max_retries` had its `_max_retries` assigned to `None`. If you wanted to use the default retry, you had to explicitly pass it a `Retry` constructed from `api.default_retry`.

If we're claiming that `api.default_retry` is a default, then it should actually be default when constructing Sessions. It's a reasonable default to use, so let's do it.

It's still possible to create a Session with a `_max_retries = None`, now you just have to do that explicitly (`None` is not the default).

In summary:

* Remove `default_retry` from `api.default_retry`, because why does anyone need it
* Make a single default `urllib3` `Retry` object to be used as the default for Sessions where the retry isn't specified
* Give `*Session` constructors the default retry (was previously no retry)
* [bonus] Allow `Sessions` to typecheck when constructed with either a `Retry` or an `int`, mirroring the interface in `requests`.
* unauthenticated sessions in the CLI explicitly have no retries
* unauthenticated session used to login explicitly has no retries
* e2e-tests explicitly create sessions with no retries when testing master for presence features
* e2e-tests use functools caching instead of self-managing it when testing master for presence of features