fix: don't access agentState when it may be nil (#8921)
AgentState is only set on the agent master-side whenever it is "started", which
is when we've made it through all the initialization messages. But an agent can
crash while it is starting, and then this access into AgentState causes a NPE,
which probably ends up being a convincing red herring when debugging.