Update notebook-entrypoint.sh
I know there was a reason I had the echo I just don't know what it was. It actually was causing issues so it is out. Added details to the unsupported architecture log line on how to skip the install. Updated where pachctl is being untarred so it doesn't show up in workdir