fix: improved textcell performance for novels (#8986)
This fixes an issue where the glide table performance would chug when
given a text cell with lots of text to truncate. while `measureText` is
more performant than getBoundingClientRectangle, it can still be
problematic to call excessively. To remedy, we implement a binary search
for the longest string that's still below max width, reducing the number
of times we need to measure the text for longer texts at the cost of
calling it a few more times for shorter texts.