refactor: consolidate experiment list selection state (#8860)
This refactors some items related to the experiment list selection ahead
of fixing the bulk action apis:

1) selection state (selected/excluded experiment ids, selectAll state)
is consolidated into a single object. Previously, selected experiment
ids, excluded experiment ids, and the selectall state were stored
separately. This is problematic -- because the selected experiment ids
were the source of truth for which experiments were selected, the stored
list would grow as the user paged through the results when in selectall
mode. we now determine the final list of selected elements on the fly.

2) grid selection is now fully determined from the id selection.
previously, the grid selection and id selection were maintained
separately. this required a lot of code ensuring they stayed in sync. we
now only derive the grid selection from the selected id list.

3) handleSelectionChange now no longer requires row ranges in an
`add-all` or `remove-all` change type. this allows for functions and
effects that solely dealt with these change types to no longer rely on
the experiments array.

4) project experiment list settings is no longer using useSettings.
because the grid selection only updates when the settings updates, the
act of updating the selection felt worse when using useSettings. we now
directly interface with the settings store at the project level --
global settings still use useSettings.

5) new hook: useTypedParams. In order to keep the functionality of the
`compare` parameter, we introduce a useTypedParams hook which allows
typesafe parsing and setting of query params similar to the settings
store api. this should allow us to finally fully migrate off of
useSettings.