fix: Experiment.await_first_trial exits when Experiment is terminal (#9022)
Pre-patch, when running await_first_trial on an Experiment the method would never exit if the experiment that backed it was in a terminal state and no trials started. It now checks that the underlying experiment is not terminal, and raises an exception if it becomes such (believing no trial will ever be started).

Side effect: Like wait, await_first_trial now reloads the underlying Experiment object.