ci: publish-docs installs awscli into user space (#9153)
PR #9126 removing the awscli installation was not well tested and doesn't work in production. This adds it back, but explicitly adds it to the user path so that the pip doesn't complain about installation into a system-managed python path.

(cherry picked from commit 0228a95b66c157b1e15e07913bfddef699593039)