fix: single-searcher MNIST example runs for multiple epochs (#9160)
When this searcher is run for a single epoch it doesn't get a chance to report progress before the entire experiment is complete. This means that the progress bar in the web jumps from 0 to 100 in a step change.

The step-change makes it look like something is broken. This eliminates that (by switching the searcher to step on batches) and demonstrates that the progress bar is working.