chore: run bulk action interstitial component
this adds an interstitial component for run bulk actions. when opened,
it runs a search given the current selection and filters and determines
whether or not the selection contains any runs that are part of a
search, returning the results to the opener. as a side effect, we pull
out the `mergeAbortControllers` helper from `trialDetailsLogs` while
also adding test coverage to that helper.