Add gen-versions.py script
Add a new Python script, gen-versions.py, to auto-generate the original,
static versions.json documentation version picker file. This script does
a few things.

First, to maintain backwards compatibility, it includes an existing
version of versions.json directly, which will be prepended to during the
release. This is because the tag-searching logic won't accurately
generate the existing file, and would probbaly include too many existing
tags that we aren't interested in (I need to check this, now that I
think about it).

Next, we use GitPython to list all tags back to a specific commit, from
all branches. We filter this list to remove release candidate tags. We
also remove the most recent tag, as that is handled specially in
versions.json.

Finally, we generate the appropriate links for each tag, including a
"latest" link for the most recent tag, and dump the entire list as JSON
to stdout, which the release process can then dump into a versions.json
file.