fix: localize debounced settings updates
This pr solves some race conditions that were added in #9220 by undoing
the global settings debounce and instead localizing it to the column
width changes the bug was talking about. There's a light refacator of
`userSettings#setPartial` which should allow it to work on any io-ts
type that has props.