fix: require full object for debounced settings updates
This fixes an issue where debounced settings slices would not update
when the settings path did not previously exist. The update function for
the debounced slice took a partial version of the slice and spread it
out with the final version of the slice. Because malformed settings
objects return null, we guarded against further issues by returning the
same if an update was made. However, unset settings also return null. To
get around this, we now require fully formed settings slices in the
debounced settings update function.