URL-encode the Determined version string
Determined master seems to know about individual agent version strings,
which are set when the agent starts, and communicated to the master via
a websocket request. The problem is that, when fetching agent
information, the version returned is incorrect when running tests. For
example, instead of "0.34.0+1c631d78f", the version string would be
returned as "0.34.0 1c631d78f", with the plus sign replaced with a
space.

This is because the agent is making a request to master with the version
string as a query parameter, which causes the master to replace the plus
sign with a space when it parses the query parameter. The solution is to
URL-encode the string so the API returns the full, literal string.

The other interpolated strings in masterAddr should probably be
URL-encoded too, but that's a larger change, and out of scope for now.