fix: prevent multiple calls to time-series on compare view select
This fixes an issue where the page could appear slower than it should on
load due to too much blocking traffic. We make three requests to build
the comparison view for runs, searches and experiments:

1) request to get search/run information of selected runs/search ids

2) request to get metric names from selected runs/searches

3) request to get timeseries data given selected run ids and metric names.

we were attempting to make these requests simultaneously when the
each request is dependent on information in the previous request.
We also used `usePrevious` hooks as dependencies of `useCallback` hooks,
which meant that we would erroneously update the callback reference when
the previous value update was picked up on subsequent renders.

this pr fixes these issues by not rendering the comparison view tabs
until the request for the records in the selection has completed, then
not loading the time series data until the metric names request has
started to return information. We also clean up some places where
effects were called due to usePrevious and the overall data flow which
should prevent issues with out of date data appearing.