fix: experiment single trial tabs don't scroll on load
This fixes an issue where loading a single trial experiment page tab
near the end of the tab list would hide the tab on mobile. We disabled
the scrolling behavior of the tab list because on certain occasions it
could scroll offpage. on removing that fix, i'm unable to reproduce the
initial issue on firefox, chrome or safari.