fix: prevent out of bounds navigation on new list views
This fixes an issue where it's possible for the user to navigate outside
of the list bounds in the new experiment view, the searches view, and
the runs view. When doing so, the user sees a blank page, which can be
confusing. In the case that the user tries to load pages beyond the
result set, we now reset the pagination back to the first page of results.