chore: re-split code editor
This fixes an issue where the code editor was being included in the main
app bundle. While you can indicate that a component is meant to be split
with `React.lazy`, if any code elsewhere imports the code statically the
code is unsplit. We wrap the code split component and ban other imports
using eslint to prevent this from regressing.