chore: deprecate SearcherContext
Instead of asking the master for SearcherOperations, the SearcherContext
fabricates a single SearcherOperation by looking at the max_length of
the experiment config.

Both the max_length of the experiment config and the SearcherContext
itself are deprecated.  If the max_length is not present, the
SearcherContext will throw an error if it is used.