INFENG-848: Make gen-versions.py match versions.sh behavior
gen-versions.py, which generates the versions.json Sphinx version picker
file that used to be managed by bumpversion, is not sensitive to where
HEAD is in the git DAG. Right now, if HEAD is on a branch that is ahead
of version 0.34.0 but behind 0.36.0, version.sh will return some flavor
of 0.35.0. gen-versions.py, however, considers all tags globally, and
will also include 0.36.0, which is incorrect.

This changeset updates gen-versions.py to be aware of where it is in the
git DAG, like version.sh does now.

* Update gen-versions.py to remove GitPython dependency and all related
logic.

* Add gen-versions.sh (for want of a better name), which returns a
newline-delimited list of all potential docs versions for
versions.json. This is basically just a single git command with some
extra filtering, along with comm to exclude certain specific versions.