chore: prune cruft out of no_op fixture
The no_op fixture is one of the oldest fixtures we have, and it has
built up a lot of cruft.

It is also based on legacy infrastructure that we are trying to remove
(SearcherContext and WorkloadSequencer), so we're first removing dead or
unnecessary code now, to make the work of rewriting the fixture easier
later.

Pointless tests, now removed:
  - test_fail_on_chechpoint_save: added in d0b384ab, used to test the
    golang workload sequencer, which is long-dead.  This is basically
    testing the fixture at this point.
  - test_fail_on_preclose_chechpoint_save: added in 2e429f30, also used
    to test the golang workload sequencer.
  - test_fail_on_first_validation: added in 39726d684ee, also used to
    test the golang workload sequencer.

Unused configs, now removed:
  - single-in-epochs.yaml
  - single-in-records.yaml
  - adaptive_chaos.yaml

Unused hyperparameters, now removed:
  - validation_secs
  - save_checkpoint_secs
  - load_checkpoint_secs
  - validation_set_size
  - fail_on_first_validation
  - fail_on_chechpoint_save

Weird cruft:
  - the metrics_sigma hparam was exclusively used to crash the trial.  I
    replaced it with the much clearer "crash_on_startup" hyperparameter.