ci: increase timeout for k8s intg tests
The latest versions of k8s have a new Job condition called
JobFailureTarget which is a signal to the jobs controller to kill off
the pods of the job.

Our jobUpdatedCallback() function waits for the JobFailure condition,
which now comes after the pod is fully terminated, which takes a lot
longer.

Probably we need to make sure our k8s logic is still valid, and deal
with the additional time it takes a job to reach JobFailed, if it
affects anything other than this test.

Until then, let's unblock CI for the whole team by just increasing the
test time for TestExternalPodDelete and TestNodeWorkflows.