fix: broken MNIST download in e2e tests
The public download URL for the MNIST dataset is slow and unreliable.

Slow is bad for CI.  So is unreliable.

So instead, when we download MNIST for the mnist_pytorch tutorial, we
include a startup hook that pre-downloads the test from an s3 bucket we
control.

S3 is fast and reliable.  That's good for CI.