ci: fix flaky test_allocation_csv tests (#9953)
test_cmd_capture was flaking around maybe 1% of the time.

I can't tell if the failure was that timestamp_before is non-inclusive
or if was just a rounding error somewhere, but adding only a couple
seconds to the boundaries seems to be enough to avoid flakes.

I fixed the other tests in the file too.  One test had been patched
already... but not the others.