fix: grid hp samping ignored empty nests
Previously, when a hyperparameter contained an emtpy "nested"
hyperparameter:

    hyperparameters:
        empty: {}

It was lost during grid sampling, and was never passed to the trial.
This is because the nested hyperparameter isn't considered its own axis
for the grid search, and is only implicitly populated by being on the
path of some inner hyperparameter axis.  However, when there is no inner
hyperparameter axis, we need to detect that and emit a one-point axis
that preserves the empty nested hyperparameter.

Another possible solution would be to convert the empty nested
hyperparameter as a const hyperparameter with an empty map for a value
when unmarshaling the hyperparameters yaml object.  I think no user code
would be affected by which solution is used, since user code will see
the same hyperparameter sample either way.