ci: remove default arg from utils.run_command()
The default arg meant a 30(!)-second sleep.

One innocent-looking test was sitting and doing nothing for a full 90
seconds simply by omitting the default arg three times.

This is a good example of why default args are bad; this default saved
somebody typing `, 0` and the cost was a test took 90 extra seconds,
probably without the author even realizing it.  Better to be clear and
explicit, especially with internal APIs (where compatibility and
breaking changes are never a concern).