fix: fix default id search for runs
this fixes an issue where the default search field for global run id
would not work properly when interacting with the filter form. the issue
is compound:
1) we defaulted new form fields to the `id` column of the experiment
location, so when searching runs the column was invalid
2) the form select uses the column name to determine what copy to show
in the select label, so the copy looked like it was using the global run
id column when it wasn't.

we fix this by specifying the location of all new children in
the form field.