fix: use UID not username to set HOME dir
When a task's uid setting is nonzero but username is still "root", we
will accidentally set the HOME directory to /root.

This happens because we start the container based on the UID, then we
look up the username (uid->username lookup) and we find that we are
called "root", then we do a username->HOME lookup to arrive at /root.