chore: remove e2e_slurm_misconfigured series tests
The only test in this series was an e2e test that the log shipper ships
a certain log in a certain situation.

We have unit tests for that, so delete the e2e test.

Now, I actually failed this e2e test when I first landed the log
shipper, so it wasn't totally good-for-nothing.  But effectively, all it
was testing these days was that we did in fact use the new log shipper
in slurm workloads.

Now that the EE code is merged into the OSS code, and we can just use
the same LogShipperWrappedEntrypoint() function in all three classes of
resource manager, I don't see any real justification for this test.

This is part of a larger effort to get rid of our znode tests, which
are notoriously unreliable.