chore: remove e2e_slurm_preemption test series
The only test in this series was an e2e test sure that we caught slurm's
SIGTERM, reported that to the master via an API call, and that training
could would then see should_preempt() return True.

This didn't need to be an e2e test; it can be a unit test to make sure
our SIGTERM handler is working, plus rerouting the API logic to pass
through the our normal preemption handling (#10071).  There is already
plenty of coverage ensuring should_preempt() works in python code.

So by adding that unit test here and following the refactor of #10071,
it is safe to remove the e2e_slurm_preemption series entirely.

This is part of a larger effort to get rid of our znode tests, which
are notoriously unreliable.