test: make explist e2e sort test more reliable
this updates the experiment list e2e multi sort test:

1) the test could flake if the table was comparing numeric values lower
than 10 with numeric values larger than 10.

2) the test helper now handles both the setup and scenario steps,
allowing for more levels of sorting.

the multi sort menu is also updated to output the values of the
direction selector to the dom to allow selections during the e2e tests
that don't need to be aware of the column copy.