chore: remove searcher context from DetCallback
The det.transformers.DetCallback chose to only work on-cluster, and
derived much of its behavior by directly inspecting ClusterInfo.

It also enforced that the user provided a transformers-style training
length which matched the searcher max_length.

And it didn't bother respecting min_validation_period or
min_checkpoint_period, or really any other training loop detail found in
the experiment config.

The end result is that with absolutely no API changes, we can actually
move to the searcher-context-removal paradigm with almost no breakages
whatsoever.

The only case that would break is if the user was _relying_ on the
asha searcher to tell the huggingface Trainer when to checkpoint.  Our
system is overall not supporting that behavior anymore, so if that
breaks it is unavoidable.