fix: remove flake-inducing logretention global singleton (#10016)
This change removes the log retention scheduler global singleton in favor of something local. The tests for log retention panicked due to a `sync.WaitGroup` going negative. After some investigation I realized we never shutdown the gocron schedulers used during testing and their tasks unconditionally call `wg.Done()`, so it's possible things we planned would only running once will run twice. An easy fix for this is to call `sched.Shutdown()` after we're done using a particular test's scheduler and an easy way to do this (mostly because the tests and code are in different packages) was to alter the package's API to return the `gocron.Scheduler` so the caller can control how shutdown happens.