ci: shorten test_pending_hpc.py (#10104)
This test contains a 60 second sleep() to ensure a log line has time to
get injected, then asserts that the log line is found.

Then it tries to kill the first experiment with a comment about speeding
up the test.  Except the underlying test is usually done in 50 seconds,
not 60.  And the log line usually appears in 10ish seconds, not 60.

So instead, wait until the log line appears, then kill the first
experiment, then wait for a RUNNING state on the second experiment, then
end the test.