build: INFENG-938: Tweak version string format (#10139)
Restore previous version tag format, which places a hyphen '-' between
the patch revision part and the remainder of the version string. So, for
example, instead of a string like v0.38.0rc3, we will look for
v0.38.0-rc3. The former is not SemVer-compliant, whereas the latteris.

This fixes an issue where Helm breaks when using release candidate
versions, because it expects the string to be
SemVer-compliant. Setuptools, by contrast, can munge the hyphenated
version into a more canonical PEP 440-compliant format, so we prefer the
new SemVer-compliant format instead.