INFENG-937: Publish Helm RCs
Add Helm makefile and CircleCI job to publish Helm chart release
candidates. This is necessary to properly deploy release candidates onto
the new release party cluster. Previously, bumpversion adjusted the
chart version in Chart.yaml and committed that to the release branch,
then rph used this in-tree version as a chart repository to do GKE
cluster upgrades during testing. This is no longer possible, though, so
we need to publish these release candidate versions. (We should probably
be doing this anyway, in the interest of transparency.)

This will create an additional GitHub release every time we cut a
release candidate.