INFENG-938: Update version format in Makefiles
Add an optional 'v' in the grep regular expression that computes
previous version tags in GoReleaser Makefile targets for master, agent,
and helm. This ensures continuity between the previous tag format, which
used bare version strings, and the new tag format, all of which begin
with a literal 'v'. Otherwise, when trying to derive a previous tag that
begins with a 'v', the grep would fail.

Note that this would work for the next release, because previous tags
would still satisfy the format. Subsequent releases, however, would be
incorrect. GORELEASER_PREVIOUS_TAG is used for determining which release
notes to include, and having it set incorrectly would cause problems for
the GitHub release.