build: INFENG-942: Conditionally bypass build-react job checks (#10145)
The release job is skipping the React build step because there are two
jobs that attempt to skip the entire job if no changes are detected in
specific directories. Previously, bumpversion made it so that, by dint
of updating various files, the job would not be skipped. Now that we
aren't using it, though, certain downstream jobs are failing because
React wasn't built.

To fix this, I've added a "force" variable to the build-react job, to
allow conditionally bypassing the checks if indicated by the
caller. This preserves existing behavior, to avoid potentially other
unintended consequences of beginning to build React everywhere, but also
to avoid running up CircleCI compute bills, as this much extra build
time would probably have a material effect on spending.