fix: client.logout() re-enables client.login()
client.login() controls a singleton, and throws an error if you try to
login twice.

But if you wanted to logout() and log back in again, that seems like it
should be allowed.

An OSS user hit this issue with a script that was meant to run for
weeks, where multiple logout() and login() calls avoided dealing with
the TTL of our sessions.