fix: cover select all edge cases
this changes the select all logic. previously the select all prompt
would appear when selecting more than the page limit's worth of objects
across the list. if the selected count equaled the total rows in the
result set, the clear selection prompt would appear. now, the select all
prompt appears when the selection includes all visible rows in the
result set, but the selection is not the entire result set. if the
selection otherwise exceeds the page limit, we show the clear selection prompt.

this fixes the select all prompt appearing when selecting single rows
across pages, as well as the select all prompt not appearing when
selecting all rows on the last page of results that does not hit the
page limit. there was also an issue where, for searches and experiments,
the select all prompt would appear when selecting 20 rows regardless of
the page size, which is also fixed here.