ci: fix incorrect field key in the goreleaser and rc label (#10254)
Fix a incorrect field key in the goreleaser config files.

Label release candidate as draft instead of pre-release. RC is
not intended for user to use. Repo viewer will not be able to see
it when it is labeled as `draft`. Taking one step back, why do
we publish rc with helm chart as an asset? There is a Github action
called `publish helm chart` publishes it to the helm repo at
helm.determined.ai. So ArgoCD can find it and upgrade the release
party cluster. For more context, check out:
https://github.com/determined-ai/determined/pull/10141